# Conflicts:
#	LegacyCubiomes/building_blocks/BasePiece.cpp
#	LegacyCubiomes/building_blocks/BasePiece.hpp
#	LegacyCubiomes/chunk_generator/CaveGenerator.hpp
#	LegacyCubiomes/chunk_generator/ChunkPrimer.hpp
#	LegacyCubiomes/chunk_generator/RavineGenerator.hpp
#	LegacyCubiomes/loot/classes/Loot.cpp
#	LegacyCubiomes/loot/classes/StrongholdLoot.hpp
#	LegacyCubiomes/loot/stronghold_corridor.hpp
#	LegacyCubiomes/structures/rolls/rolls_base.cpp
#	LegacyCubiomes/structures/rolls/rolls_base.hpp
#	LegacyCubiomes/structures/structure_rolls/stronghold_rolls.hpp
#	main1.cpp

+++ LegacyCubiomes/enchants/classes/EnchantmentArrowDamage.hpp
@@ -5,8 +5,8 @@
 
 class EnchantmentArrowDamage : public Enchantment {
 private:
-    static constexpr inline int LEVELS_MIN[5] = {9, 19, 29, 39, 49};
-    static constexpr inline int LEVELS_MAX[5] = {16, 26, 36, 46, 56};
+    static constexpr inline int8_t LEVELS_MIN[5] = {9, 19, 29, 39, 49};
+    static constexpr inline int8_t LEVELS_MAX[5] = {16, 26, 36, 46, 56};
 public:
 
     EnchantmentArrowDamage(const std::string& name, const Rarity *rarity) :
