+++ LegacyCubiomes/enchants/classes/EnchantmentArrowKnockback.hpp
@@ -5,8 +5,8 @@
 
 class EnchantmentArrowKnockback : public Enchantment {
 private:
-    static constexpr inline int LEVELS_MIN[2] = {12, 32};
-    static constexpr inline int LEVELS_MAX[2] = {37, 57};
+    static constexpr inline int8_t LEVELS_MIN[2] = {12, 32};
+    static constexpr inline int8_t LEVELS_MAX[2] = {37, 57};
 
 public:
