+++ LegacyCubiomes/enchants/classes/EnchantmentDamage.hpp
@@ -7,9 +7,9 @@ class EnchantmentDamage : public Enchantment {
 public:
 
     const int damageType;
-    const std::vector<int> BASE_ENCHANTABILITY  = {1, 5, 5};
-    const std::vector<int> LEVEL_ENCHANTABILITY  = {11, 8, 8};
-    const std::vector<int> THRESHOLD_ENCHANTABILITY  = {20, 20, 20};
+    const std::vector<int8_t> BASE_ENCHANTABILITY  = {1, 5, 5};
+    const std::vector<int8_t> LEVEL_ENCHANTABILITY  = {11, 8, 8};
+    const std::vector<int8_t> THRESHOLD_ENCHANTABILITY  = {20, 20, 20};
 
 
     EnchantmentDamage(std::string name, const Rarity *rarity, int damage_type) :
