+++ LegacyCubiomes/enchants/classes/EnchantmentWaterWalker.hpp
@@ -5,8 +5,8 @@
 
 class EnchantmentWaterWalker : public Enchantment {
 private:
-    static constexpr inline int LEVELS_MIN[3] = {10, 20, 30};
-    static constexpr inline int LEVELS_MAX[3] = {25, 35, 45};
+    static constexpr inline int8_t LEVELS_MIN[3] = {10, 20, 30};
+    static constexpr inline int8_t LEVELS_MAX[3] = {25, 35, 45};
 public:
 
     EnchantmentWaterWalker(std::string name, const Rarity *rarity) :
